ML_model/requirements.txt
onnxruntime==1.18.1
numpy
Pillow
